[FIX/#143] cacheDir File Output 뷰모델 로직 이동
