+++ presentation/src/main/java/kr/genti/presentation/generate/finished/FinishedViewModel.kt
@@ -1,17 +1,24 @@
 package kr.genti.presentation.generate.finished
 
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kr.genti.domain.entity.request.ReportRequestModel
 import kr.genti.domain.entity.response.ImageModel
 import kr.genti.domain.enums.PictureRatio
 import kr.genti.domain.repository.GenerateRepository
 import kr.genti.domain.repository.UserRepository
+import java.io.File
+import java.io.FileOutputStream
+import java.net.URL
 import javax.inject.Inject
 
 @HiltViewModel
@@ -35,6 +42,9 @@ class FinishedViewModel
 
         var isRatioGaro = true
 
+        private val _isImageDownloaded = MutableSharedFlow<Boolean>()
+        val isImageDownloaded: SharedFlow<Boolean> = _isImageDownloaded
+
         private val _postReportResult = MutableSharedFlow<Boolean>()
         val postReportResult: SharedFlow<Boolean> = _postReportResult
 
@@ -50,6 +60,27 @@ class FinishedViewModel
 
         fun setPictureRatio() {
             isRatioGaro = finishedImage.pictureRatio == PictureRatio.RATIO_GARO
+        }
+
+        fun downloadImageToCache(outputFile: File) {
+            viewModelScope.launch {
+                runCatching {
+                    withContext(Dispatchers.IO) {
+                        val connection = URL(finishedImage.url).openConnection()
+                        connection.connect()
+                        connection.getInputStream()?.use { inputStream ->
+                            FileOutputStream(outputFile).use { out ->
+                                val bitmap = BitmapFactory.decodeStream(inputStream)
+                                bitmap?.compress(Bitmap.CompressFormat.PNG, 100, out)
+                            }
+                        }
+                    }
+                }.onSuccess {
+                    _isImageDownloaded.emit(true)
+                }.onFailure {
+                    _isImageDownloaded.emit(false)
+                }
+            }
         }
 
         fun postGenerateRateToServer(star: Int) {
